refactor: second-pass simplification — dead code, dedup, perf, correctness (#144)
Dead code removal:
- Remove 10 unused functions/methods/constants across hotkey.py, config.py,
  result_window_web.py, vocab_build_window.py, recording_flow.py,
  clipboard_monitor.py, apple.py

Deduplication:
- Extract shared close delegate factory in web_utils.py (7 panels unified)
- Extract _normalize_key_name() in hotkey.py (7 sites unified)
- Extract _resize_icon_to_png() in file_source.py (2 functions unified)
- Extract _rebuild_content_view() in recording_indicator.py
- Extract _rewrite_matching_record() in conversation_history.py

Performance:
- Use int32 instead of float64 for RMS in audio callback (50x/sec)
- Cache running app names with 2s TTL in app_source.py
- Cache bookmark domain on first access in bookmark_source.py

Correctness:
- Replace assert with ValueError in sherpa.py (survives -O flag)
- Replace deprecated lockFocus with NSGraphicsContext in app_source.py
  and file_source.py (thread-safe, not deprecated)

Net: -154 lines across 21 files. All 3,684 tests pass.

+++ src/wenzi/audio/recorder.py
@@ -244,7 +244,7 @@ class Recorder:
 
         audio = np.concatenate(frames)
         duration = len(audio) / self.sample_rate
-        rms = int(np.sqrt(np.mean(audio.astype(np.float64) ** 2)))
+        rms = int(np.sqrt(np.mean(audio.astype(np.int32) ** 2)))
         logger.info(
             "Recording stopped, captured %d samples (%.1fs), RMS=%d",
             len(audio), duration, rms,
@@ -286,7 +286,7 @@ class Recorder:
             logger.warning("Max session size reached, dropping frames")
             return
 
-        self._current_rms = float(np.sqrt(np.mean(frame.astype(np.float64) ** 2)))
+        self._current_rms = float(np.sqrt(np.mean(frame.astype(np.int32) ** 2)))
         self._total_bytes += frame_bytes
         copied = frame.copy()
         try:
